feat: 좋아요 기능 구현 & add: PostCard.jsx 에 좋아요 해제 로직 추가

+++ src/services/likeService.js
@@ -4,20 +4,11 @@ import supabase from "../supabaseClient";
 const updateLikeStatus = async (post_id, user_id, isLike) => {
   try {
     if (!isLike) {
-      const { data } = await supabase
-        .from("likes")
-        .insert({ post_id: post_id, user_id: user_id });
-
-      console.log("좋아요 추가");
+      const { data } = await supabase.from("likes").insert({ post_id: post_id, user_id: user_id });
     } else {
-      const { data, error } = await supabase
-        .from("likes")
-        .delete()
-        .eq("post_id", post_id)
-        .eq("user_id", user_id);
+      const { data, error } = await supabase.from("likes").delete().eq("post_id", post_id).eq("user_id", user_id);
 
       if (error) throw error;
-      console.log("좋아요 삭제");
     }
   } catch (error) {
     console.error("좋아요 업데이트에 실패했습니다.", error);
